Move more app files to app package
