+++ tests/unit/test_spinner.py
@@ -1,7 +1,7 @@
 # Generated by CodiumAI
 import time
 
-from autogpt.spinner import Spinner
+from autogpt.app.spinner import Spinner
 
 """
 Code Analysis
